moving _calculate_distance
