+++ landlab/components/profiler/channel_profiler.py
@@ -4,6 +4,7 @@
 
 import numpy as np
 
+from landlab import RasterModelGrid
 from landlab.components.profiler.base_profiler import _BaseProfiler
 
 
@@ -420,3 +421,42 @@ class ChannelProfiler(_BaseProfiler):
                     channel_network.append(np.array(channel_segment))
                     queue.extend(nodes_to_process)
                 self._profile_structure.append(channel_network)
+
+    def _calculate_distances(self):
+        """Get distances along the profile_IDs datastructure."""
+        self._distance_along_profile = []
+        end_distances = {}
+
+        # set the starting values for the beginnings of each netwrok.
+        for network in self._profile_structure:
+            starting_node = network[0][0]
+            end_distances[starting_node] = 0
+
+        # for each network
+        for network in self._profile_structure:
+
+            network_values = []
+            # for each segment in the network.
+            for segment in network:
+                starting_node = segment[0]
+
+                total_distance = end_distances[starting_node]
+
+                profile_values = []
+                profile_values.append(total_distance)
+
+                # itterate up the profile
+                for j in range(len(segment) - 1):
+                    if isinstance(self._grid, RasterModelGrid):
+                        total_distance += self._grid.length_of_d8[
+                            self._link_to_flow_receiver[segment[j + 1]]
+                        ]
+                        profile_values.append(total_distance)
+                    else:
+                        total_distance += self._grid.length_of_link[
+                            self._link_to_flow_receiver[segment[j + 1]]
+                        ]
+                        profile_values.append(total_distance)
+                network_values.append(np.array(profile_values))
+                end_distances[segment[-1]] = total_distance
+            self._distance_along_profile.append(network_values)
